pyproject.toml
[project]
name = "langchain-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=25.1.0",
    "dotenv>=0.9.9",
    "isort>=6.0.1",
    "langchain>=0.3.27",
    "langchain-ollama>=0.3.6",
    "langchain-openai>=1.1.7",
    "langchain-tavily>=0.2.17",
    "python-dotenv>=1.1.1",
    "tavily-python>=0.7.19",
]
